# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Next Olshop API",
    "description": "REST API for Next Olshop — Single Store Online Shop untuk UMKM. Digunakan oleh Flutter mobile app.\n\n## Authentication\nSemua endpoint yang membutuhkan auth menggunakan Bearer token dari `/api/auth/login`.\n\n## Response Format\n- **Success**: `{ success: true, data: T }`\n- **Error**: `{ success: false, error: string }`\n- **Pagination**: `{ success: true, data: T[], meta: { page, limit, total, totalPages } }`\n\n## Catatan\n- Harga dalam Rupiah (IDR), disimpan sebagai string decimal (`\"299000.00\"`)\n- Tanggal dalam format ISO 8601 (UTC)\n- Berat produk dalam gram",
    "version": "1.0.0",
    "contact": {
      "name": "Next Olshop",
      "url": "https://github.com/codewithbahri"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token dari POST /api/auth/login"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 2 },
          "name": { "type": "string", "example": "John Doe" },
          "email": { "type": "string", "format": "email", "example": "john@example.com" },
          "phone": { "type": "string", "nullable": true, "example": "08123456789" },
          "role": { "type": "string", "enum": ["customer", "admin"], "example": "customer" }
        }
      },
...
      "CartItem": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "userId": { "type": "integer", "example": 2 },
          "productId": { "type": "integer", "example": 1 },
          "quantity": { "type": "integer", "example": 2 },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "product": {
            "type": "object",
            "properties": {
              "id": { "type": "integer" },
              "name": { "type": "string" },
              "slug": { "type": "string" },
              "price": { "type": "string", "example": "299000.00" },
              "image": { "type": "string", "nullable": true },
              "stock": { "type": "integer" },
              "weight": { "type": "integer" },
              "isActive": { "type": "boolean" },
              "isFeatured": { "type": "boolean" }
            }
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "Error message" }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "Validasi gagal" },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": { "type": "string" }
            },
            "example": { "email": ["Email tidak valid"], "password": ["Password minimal 6 karakter"] }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register",
        "description": "Daftar akun customer baru",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string", "example": "John Doe" },
                  "email": { "type": "string", "format": "email", "example": "john@example.com" },
                  "password": { "type": "string", "minLength": 6, "example": "password123" },
                  "phone": { "type": "string", "example": "08123456789" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registrasi berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "name": { "type": "string" },
                        "email": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error atau email sudah terdaftar",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get Profile",
        "description": "Ambil data user yang sedang login",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
    "/api/cart": {
      "get": {
        "tags": ["Cart"],
        "summary": "Get Cart",
        "description": "Ambil isi keranjang user beserta detail produk",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Cart items with product",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/CartItem" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      },
      "post": {
        "tags": ["Cart"],
        "summary": "Add to Cart",
        "description": "Tambah produk ke keranjang. Jika produk sudah ada, quantity ditambah.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["productId"],
                "properties": {
                  "productId": { "type": "integer", "example": 1 },
                  "quantity": { "type": "integer", "default": 1, "example": 2 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item ditambahkan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "quantity": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Stok tidak mencukupi",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Produk tidak ditemukan",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/api/cart/{id}": {
      "put": {
        "tags": ["Cart"],
        "summary": "Update Cart Item",
        "description": "Update quantity item di keranjang",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "description": "Cart item ID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["quantity"],
                "properties": {
                  "quantity": { "type": "integer", "minimum": 1, "example": 3 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Quantity updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "quantity": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Item tidak ditemukan",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      },
      "delete": {
        "tags": ["Cart"],
        "summary": "Remove Cart Item",
        "description": "Hapus item dari keranjang",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "description": "Cart item ID" }
        ],
        "responses": {
          "200": {
            "description": "Item dihapus",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Item tidak ditemukan",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
  }
}